requirements.txt
Django==5.0.2
channels[daphne]==4.0.0
channels-redis==4.1.0
psycopg2-binary==2.9.9
django-htmx==1.17.0
Pillow==10.2.0
mss==9.0.1
pynput==1.7.6
requests==2.31.0
netifaces==0.11.0
psutil==5.9.8
cryptography==42.0.5